Changes while trying to fix debian packaging issues (#1372)
* packaging: derive version from package.xml; add missing optional deps

- Remove version_prefix.txt: its version was stale (2.15.x) and unused by
  CMake, which already reads the version from modules/mrpt_common/package.xml.
- parse_mrpt_version.sh now extracts the version from mrpt_common/package.xml
  (robust multi-digit major/minor/patch) and is the single source of truth;
  make_release.sh and doc/Makefile use it.
- Switch root-directory sentinel checks from version_prefix.txt to
  colcon_defaults.yaml in helper scripts.
- Remove obsolete scripts/bootstrap_new_lib.py (referenced the old libs/ layout).
- Add optional auto-detected build deps to package.xml so downstream packaging
  (Debian) picks them up: ffmpeg (avcodec/avformat/avutil/swscale), libdc1394,
  openni2, libpcap, libusb (mrpt_hwdrivers) and libftdi1 (mrpt_comms).

cmake: allow overriding python install dir via MRPT_PYTHON_INSTALL_DIR

Debian packaging passes -DMRPT_PYTHON_INSTALL_DIR=lib/python3/dist-packages
to install pybind11 modules to the Debian-standard dist-packages location
instead of the colcon/ROS site-packages default.

package.xml: declare required system deps tinyxml2 (obs) and octomap (maps)

These are detected at build time and, if the system version is absent, the
CMake scripts fall back to downloading/building an embedded copy. Declaring
them lets distro packaging (Debian) provide the system versions and avoid
network access during the build.

containers: include public <libfyaml.h> instead of internal header

yaml.cpp included <libfyaml/libfyaml-core.h>, an internal libfyaml header
only present in the vendored source tree. Distro packages (e.g. Debian
libfyaml-dev) ship only the public umbrella header <libfyaml.h>, which works
for both the system and embedded builds. This restores the ability to build
against the system libfyaml (as MRPT 2.x did).

package.xml: declare system deps for simpleini+icu, nanoflann, libfyaml

These libraries are detected from the system (with the vendored copies as
fallback). Declaring them lets distro packaging install the system versions:
- mrpt_config: libsimpleini-dev + libicu-dev (SI_CONVERT_ICU)
- mrpt_math: libnanoflann-dev
- mrpt_containers: libfyaml-dev

hwdrivers: fix CFFMPEG_InputStream build with modern FFmpeg

avcodec_close() was deprecated and removed in recent FFmpeg (libavcodec 61,
Debian sid). For the alloc'd codec context path (LIBAVFORMAT_VERSION_MAJOR>=57)
use avcodec_free_context(), which both closes and frees it; keep avcodec_close()
only on the legacy (<57) path.

* update libfyaml submodule

* fix: point libfyaml submodule to fork with C++ atomic-macro-leak fix

GCC C++ builds of mrpt_containers fail because libfyaml-atomics.h's
GCC-in-C++ fallback path leaves unprefixed atomic_* macros defined,
which corrupt libstdc++'s <bits/shared_ptr_atomic.h> once <memory> is
included later in the same translation unit (e.g. in yaml.cpp).

Pin to jlblancoc/libfyaml's fix/cxx-atomic-macro-leak-shared-ptr branch
until upstream PR https://github.com/pantoniou/libfyaml/pull/310 is
merged, then switch .gitmodules back to pantoniou/libfyaml.

* fix: update libfyaml fork pin to corrected atomic-macro-leak fix

The first fix attempt (#undef-ing colliding atomic_* macros at the end
of libfyaml-atomics.h) broke libfyaml's own C++ atomics test on every
platform in upstream CI, since fy_atomic_*() legitimately expands to
those names after the header is processed. Replaced with a rename-based
fix (fy_priv_atomic_* private namespace) that avoids the collision
without ever undefining anything. See upstream PR discussion:
https://github.com/pantoniou/libfyaml/pull/310

* fix(hwdrivers): use LIBAVCODEC_VERSION_MAJOR to gate avcodec_free_context()

avcodec_free_context() is part of libavcodec, not libavformat, and was
introduced in libavcodec 55.52.0. The previous check against
LIBAVFORMAT_VERSION_MAJOR compared the wrong library's version. Also
add the missing av_free() on the legacy path: avcodec_close() only
releases internal codec data, not the AVCodecContext struct itself.

Addresses coderabbitai review comment on MRPT/mrpt#1372.

* fix: update libfyaml fork pin to fix Xcode 16/macOS 15 <atomic> error

macos-15-xcode-16 CI was failing with:
  error: <atomic> is incompatible with <stdatomic.h> before C++23.
because libfyaml.h pulled in the real <stdatomic.h> for Clang, then
yaml.cpp transitively includes libc++'s <atomic> via <optional>.
Recent libc++ hard-errors on that combination pre-C++23. Fixed
upstream by restricting <stdatomic.h> use in C++ mode to C++23+; see
https://github.com/pantoniou/libfyaml/pull/310 for details.

* fix: update libfyaml fork pin to fix Clang _Atomic/__c11_atomic_* bug

The previous pin re-enabled Clang's native _Atomic qualifier (needed so
libc++'s own <atomic> implementation, which relies on it internally,
isn't corrupted by our no-op fallback) but didn't account for Clang's
private (non-stdatomic.h) atomic ops needing real __c11_atomic_*
compiler builtins rather than the GCC-style __typeof__ fallback.
Verified locally with gcc and clang-18 (test/libfyaml-test-cpp-atomics.cpp
compiles and runs correctly under both) and MRPT's mrpt_containers
test suite (56/56 passing). See
https://github.com/pantoniou/libfyaml/pull/310 for full details.

* fix: update libfyaml fork pin, exclude Windows from Clang _Atomic path

CI on the upstream PR caught a Clang-on-Windows-Debug-specific codegen
bug in __c11_atomic_fetch_add() (wrong return value, both windows-2022
and windows-latest). Windows Clang links against MSVC's STL, not
libc++, so it never needed the real-_Atomic/__c11_atomic_* path to
begin with; it now falls through to the same safe non-atomic fallback
GCC uses. See https://github.com/pantoniou/libfyaml/pull/310 for the
full story.

* fix(containers): isolate libfyaml.h behind a plain-C shim, drop the fork

yaml.cpp included <libfyaml.h> directly from C++, which repeatedly broke
across compilers (Clang/_Atomic on Windows, Xcode 16 <atomic>, etc.) due
to the library's internal atomics shim only misbehaving in C++ mode.
Add yaml_fy_shim.{c,h}, a POD-only event API compiled as plain C, and
have yaml.cpp talk to libfyaml only through it. No fy_* type crosses
the C/C++ boundary anymore, so the whole bug class no longer applies
to MRPT regardless of upstream/compiler quirks.

With that isolation in place, the local libfyaml fork (carrying five
C++-interop patches) is no longer needed; point the submodule back at
upstream pantoniou/libfyaml.

+++ mrpt_examples_cpp/slam_icp3d_simple_example/main.cpp
@@ -60,8 +60,9 @@ void generatePointClouds(CSimplePointsMap& M1, CSimplePointsMap& M2)
     const double phi = std::acos(rng.drawUniform(-1.0, 1.0));
     const double r = 0.5;
     M1.insertPoint(
-        r * std::sin(phi) * std::cos(theta), r * std::sin(phi) * std::sin(theta),
-        r * std::cos(phi));
+        static_cast<float>(r * std::sin(phi) * std::cos(theta)),
+        static_cast<float>(r * std::sin(phi) * std::sin(theta)),
+        static_cast<float>(r * std::cos(phi)));
   }
 
   // Disk 1: radius 2, pose (0,0,0, 0, 5deg, 5deg)
@@ -73,7 +74,8 @@ void generatePointClouds(CSimplePointsMap& M1, CSimplePointsMap& M2)
     mrpt::math::TPoint3D local(r * std::cos(ang), r * std::sin(ang), 0);
     mrpt::math::TPoint3D world;
     disk1_pose.composePoint(local, world);
-    M1.insertPoint(world.x, world.y, world.z);
+    M1.insertPoint(
+        static_cast<float>(world.x), static_cast<float>(world.y), static_cast<float>(world.z));
   }
 
   // Disk 2: radius 2, pose (0, 0, 0, 30deg, -20deg, -2deg)
@@ -85,7 +87,8 @@ void generatePointClouds(CSimplePointsMap& M1, CSimplePointsMap& M2)
     mrpt::math::TPoint3D local(r * std::cos(ang), r * std::sin(ang), 0);
     mrpt::math::TPoint3D world;
     disk2_pose.composePoint(local, world);
-    M1.insertPoint(world.x, world.y, world.z);
+    M1.insertPoint(
+        static_cast<float>(world.x), static_cast<float>(world.y), static_cast<float>(world.z));
   }
 
   // M2 is M1 with a known pose error applied
